refactor: switch current code to typescript
